pyproject.toml
[tool.poetry]
name = "pyecs"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
typing-extensions = "^4.8.0"
starlette = "^0.31.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
adaptix = "^3.0.0a5"
attrs = "^23.1.0"
anyio = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
ruff = "^0.0.292"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
src = ["game", "tests"]

[tool.isort]
profile = "pycharm"
line_length = 90